+++ test/tempo/local_date_test.dart
@@ -50,9 +50,32 @@ void main() {
     });
 
     test('now()', () {
-      var d =
-          withClock(Clock.fixed(DateTime(2000, 1, 2)), () => LocalDate.now());
-      expect(d, hasDate(2000, 1, 2));
+      // This is a bit tricky. We need to ensure that now() tracks
+      // defaultZoneId and not whatever time zone DateTime happens to be using.
+      // (Which, unfortunately, can't be portably discovered in Dart.)
+      //
+      // Here's what's happening:
+      //
+      //   - Set defaultZoneId to a value that's unlikely* to be DateTime's
+      //     default
+      //   - Set the clock to midnight in that unlikely timezone
+      //   - Test each hour of the day
+      //
+      // If DateTime's time zone doesn't match, then at least some of those
+      // hours will occur on the wrong day, and the test will fail.
+      //
+      // * If you're running this test in the easternmost islands of Kiribati,
+      // hello! I apologize. This test won't be quite as effective, but it
+      // should at least pass.
+      defaultZoneId = 'Pacific/Kiritimati';
+      for (var hour = 0; hour <= 23; hour++) {
+        var d = withClock(
+            Clock.fixed(DateTime.utc(2026, 1, 2, hour)
+                .subtract(Duration(hours: 14)) // Kiritimati's UTC offset
+                .toLocal()),
+            () => LocalDate.now());
+        expect(d, hasDate(2026, 1, 2));
+      }
     });
   });
 
@@ -221,8 +244,8 @@ void main() {
     });
 
     test('inTimezone', () {
-      expect(date.inTimezone(), hasUnixSeconds(unixSeconds));
-      expect(date.inTimezone().zoneId, defaultZoneId);
+      final want = ZonedDateTime(2000, 1, 2);
+      expect(date.inTimezone(), want);
     });
 
     test('inTimezone with time zone', () {